+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html) and
 ### Removed
 
 - Deprecated methods
+- `SmartEnum` (use `CleverEnum` instead)
 
 ### Fixed